+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Additions
 - [Issue #189] Add CanReadAll() and CanWriteAll()
 
+### Fixes
+- npm update
+
 ## [3.1.0] Aug-06-2019
 
 ### Additions